+++ src/main/java/com/devnity/devnity/common/utils/s3/HealthController.java
@@ -1,4 +1,4 @@
-package com.devnity.devnity.common.utils;
+package com.devnity.devnity.common.utils.s3;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
