+++ src/main/java/com/devnity/devnity/common/utils/validator/PasswordValidator.java
@@ -1,6 +1,6 @@
-package com.devnity.devnity.domain.user.utils.validator;
+package com.devnity.devnity.common.utils.validator;
 
-import com.devnity.devnity.domain.user.utils.annotation.Password;
+import com.devnity.devnity.common.utils.annotation.Password;
 import java.util.Objects;
 import java.util.regex.Pattern;
 import javax.validation.ConstraintValidator;
